functional tests only cleanup if succeeded

+++ tests/functional/test_ROSE_ORIG_HOST.py
@@ -84,7 +84,7 @@ def monkeymodule():
 
 
 @pytest.fixture(scope='module')
-def fixture_provide_flow(tmp_path_factory):
+def fixture_provide_flow(tmp_path_factory, request):
     """Provide a cylc workflow based on the contents of a folder which can
     be either validated or installed.
     """
@@ -99,8 +99,9 @@ def fixture_provide_flow(tmp_path_factory):
         'flowpath': flowpath,
         'srcpath': srcpath
     }
-    shutil.rmtree(srcpath)
-    shutil.rmtree(flowpath)
+    if not request.session.testsfailed:
+        shutil.rmtree(srcpath)
+        shutil.rmtree(flowpath)
 
 
 @pytest.fixture(scope='module')
